% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModelTest.r
\name{linearModelTest}
\alias{linearModelTest}
\title{A test function for linear models}
\usage{
linearModelTest(n, p, beta.not.null = c(1, 2, 3), num.select = 5 *
  length(beta.not.null), ebic = FALSE)
}
\arguments{
\item{n}{the sampel size}

\item{p}{the model dimension}

\item{beta.not.null}{the non-zero coefficient indexes}

\item{num.select}{the same as in "screening"}

\item{ebic}{the same as in "screening"}
}
\description{
This function tests screening performance on linear models with user specified dimension. The function generates synthetic data using linear model and prints the screened variable for all method to the screen.
}

