\name{myscreening-package}
\alias{myscreening-package}
\alias{myscreening}
\docType{package}
\title{
\packageTitle{myscreening}
}
\description{
\packageDescription{myscreening}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myscreening}
\packageIndices{myscreening}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{myscreening}

Maintainer: \packageMaintainer{myscreening}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
