% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Code to use the screening package inline. Not directly called by the user.}
\usage{
inlineCxxPlugin(...)
}
\arguments{
\item{...}{arguments}
}
\description{
Code to use the screening package inline. Not directly called by the user.
}
\keyword{internal}
