% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screening_model_1_likes}
\alias{screening_model_1_likes}
\title{Do likelihood calculations for ScreeningModel1}
\usage{
screening_model_1_likes(
  inputs,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta = 0.05
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{scale1}{Weibull scale for onset}

\item{shape2}{Weibull shape for clinical diagnosis}

\item{scale2}{Weibull scale for clinical diagnosis}

\item{beta}{false negative fraction for screening}

\item{tj}{double vector of screening times}

\item{shape}{Weibull shape for onset}
}
\value{
list with element t for the times t and element p for a numeric matrix of the probabilities, including column names X, Y and Z.
}
\description{
Do likelihood calculations for ScreeningModel1
}
