% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screening_model_1_likes_MVK_exp_grad}
\alias{screening_model_1_likes_MVK_exp_grad}
\title{Do AD likelihood and gradient calculations for ScreeningModel1 (MVK + Exp)}
\usage{
screening_model_1_likes_MVK_exp_grad(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  rate = 0.1,
  beta = 0.05,
  tol = 1e-06,
  n_threads = 0L
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{A}{MVK parameter A (active for AD)}

\item{B}{MVK parameter B (active for AD)}

\item{delta}{MVK parameter delta (active for AD)}

\item{rate}{Exponential rate for clinical diagnosis (active for AD)}

\item{beta}{false negative fraction for screening (active for AD)}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}

\item{n_threads}{number of threads to use (default=0, auto-detects)}
}
\value{
list containing likelihoods and gradients
}
\description{
Do AD likelihood and gradient calculations for ScreeningModel1 (MVK + Exp)
}
