% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screening_model_1_predictions}
\alias{screening_model_1_predictions}
\title{Do predictions for ScreeningModel1}
\usage{
screening_model_1_predictions(
  time,
  tj,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta = 0.05
)
}
\arguments{
\item{tj}{double vector of screening times}

\item{scale1}{Weibull scale for onset}

\item{shape2}{Weibull shape for clinical diagnosis}

\item{scale2}{Weibull scale for clinical diagnosis}

\item{beta}{false negative fraction for screening}

\item{t}{double vector of times to evaluate}

\item{shape}{Weibull shape for onset}
}
\value{
data-frame with elements time, X, Y, Z (for state probabilities) and I (for incidence)
}
\description{
Do predictions for ScreeningModel1
}
\examples{
par(mfrow=1:2)
x=seq(0,30,length=301)
screening_model_1_predictions(t=x, tj=vector("double"), shape1=1.5, shape2=1.5, scale1=20, scale2=10, beta=0.4) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="No screening"))
screening_model_1_predictions(t=x, tj=c(10,15, 20, 25), shape1=1.5, shape2=1.5, scale1=20, scale2=10, beta=0.4) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="Screening"))
}
