% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screening_model_4_likes_loglin_grad}
\alias{screening_model_4_likes_loglin_grad}
\title{Do AD likelihood and gradient calculations for ScreeningModel4
(MVK onset + Exp sojourn + log-linear PSA with slope change after onset)}
\usage{
screening_model_4_likes_loglin_grad(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  rate = 0.1,
  beta0 = 3.2892,
  beta1 = -0.5533,
  b0_psa = -1.6094,
  b1_psa = 0.02,
  b2_psa = 0.1094,
  sigma_psa = 0.2879,
  PrFalseNegBx = 0,
  tol = 1e-06,
  n_threads = 0L
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time,
ti for screening times, yi for biomarker values, bxi for biopsy indicators,
and type for the type of likelihood
(1 = no cancer detected, 2 = screen-detected cancer, 3 = interval cancer)}

\item{A}{MVK parameter A (active for AD)}

\item{B}{MVK parameter B (active for AD)}

\item{delta}{MVK parameter delta (active for AD)}

\item{rate}{Exponential rate for clinical diagnosis (active for AD)}

\item{beta0}{intercept for logistic no-biopsy model (active for AD)}

\item{beta1}{slope of log(yi) for logistic no-biopsy model (active for AD)}

\item{b0_psa}{intercept for log-linear PSA model (active for AD)}

\item{b1_psa}{age slope for log-linear PSA model (active for AD)}

\item{b2_psa}{slope increment after onset for log-linear PSA model (active for AD)}

\item{sigma_psa}{standard deviation of log(PSA) (active for AD)}

\item{PrFalseNegBx}{probability of a false negative biopsy | cancer, biopsy undertaken (active for AD)}

\item{tol}{double for numeric integration tolerance}

\item{n_threads}{number of threads to use (default = 0, auto-detects)}
}
\value{
list containing likelihoods and gradients
}
\description{
Do AD likelihood and gradient calculations for ScreeningModel4
(MVK onset + Exp sojourn + log-linear PSA with slope change after onset)
}
