% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screening_model_5_likes_loglin_grad}
\alias{screening_model_5_likes_loglin_grad}
\title{Do AD likelihood and gradient calculations for ScreeningModel5
(MVK onset + Exp sojourn + log-linear PSA + random effects for intercept)}
\usage{
screening_model_5_likes_loglin_grad(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  rate = 0.1,
  beta0 = 3.2892,
  beta1 = -0.5533,
  mu_b0 = -1.6094,
  sigma_b0 = 0.2383,
  b1_psa = 0.02,
  b2_psa = 0.1094,
  sigma_psa = 0.2879,
  PrFalseNegBx = 0,
  tol = 1e-06,
  gh_nodes = NULL,
  gh_weights = NULL,
  n_threads = 0L
)
}
\arguments{
\item{inputs}{list of list with elements of t, type, ti, yi, and bxi}

\item{A}{MVK parameter A (active for AD)}

\item{B}{MVK parameter B (active for AD)}

\item{delta}{MVK parameter delta (active for AD)}

\item{rate}{Exponential rate for clinical diagnosis (active for AD)}

\item{beta0}{intercept for logistic no-biopsy model (active for AD)}

\item{beta1}{slope of log(yi) for logistic no-biopsy model (active for AD)}

\item{mu_b0}{mean intercept for log-linear PSA model (active for AD)}

\item{sigma_b0}{standard deviation of intercept for log-linear PSA model (active for AD)}

\item{b1_psa}{age slope for log-linear PSA model (active for AD)}

\item{b2_psa}{slope increment after onset for log-linear PSA model (active for AD)}

\item{sigma_psa}{standard deviation of log(PSA) (active for AD)}

\item{PrFalseNegBx}{probability of a false negative biopsy (active for AD)}

\item{tol}{double for numeric integration tolerance}

\item{gh_nodes}{vector of Gauss-Hermite nodes}

\item{gh_weights}{vector of Gauss-Hermite weights}

\item{n_threads}{number of threads to use (default = 0, auto-detects)}
}
\value{
list containing likelihoods and gradients
}
\description{
Do AD likelihood and gradient calculations for ScreeningModel5
(MVK onset + Exp sojourn + log-linear PSA + random effects for intercept)
}
