% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScreeningModel1}
\alias{ScreeningModel1}
\alias{screening_model_1_predictions}
\alias{screening_model_1_likes}
\title{Do predictions for ScreeningModel1}
\usage{
screening_model_1_predictions(
  t,
  ti,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta = 0.05,
  simple = TRUE,
  tol = 1e-06
)

screening_model_1_likes(
  inputs,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta = 0.05,
  tol = 1e-06
)
}
\arguments{
\item{t}{double vector of times to evaluate}

\item{ti}{double vector of screening times}

\item{shape1}{Weibull shape for onset}

\item{scale1}{Weibull scale for onset}

\item{shape2}{Weibull shape for clinical diagnosis}

\item{scale2}{Weibull scale for clinical diagnosis}

\item{beta}{false negative fraction for screening}

\item{simple}{whether to calculate Z using a simple method (complement) or integration (defaults to true)}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}

\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}
}
\value{
data-frame with elements t, X, Y, Z (for state probabilities) and I (for incidence)

vector of likelihoods
}
\description{
Do predictions for ScreeningModel1

Do likelihood calculations for ScreeningModel1
}
\examples{
par(mfrow=1:2)
x=seq(0,30,length=301)
screening_model_1_predictions(t=x, tj=vector("double"), shape1=1.5, shape2=1.5, scale1=20, scale2=10, beta=0.4) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="No screening"))
screening_model_1_predictions(t=x, tj=c(10,15, 20, 25), shape1=1.5, shape2=1.5, scale1=20, scale2=10, beta=0.4) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="Screening"))
}
