% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScreeningModel2MVKExp}
\alias{ScreeningModel2MVKExp}
\alias{screening_model_2_likes_MVK_exp}
\title{Do likelihood calculations for ScreeningModel2 using MVK onset and Exponential Sojourn}
\usage{
screening_model_2_likes_MVK_exp(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  rate = 0.1,
  beta0 = -3,
  beta1 = 1,
  tol = 1e-06
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{A}{MVK parameter A (typically negative, related to net proliferation)}

\item{B}{MVK parameter B (typically positive, related to malignant transformation)}

\item{delta}{MVK parameter delta (ratio of initiation rate to cell division rate)}

\item{rate}{Exponential rate for clinical diagnosis (Sojourn density, default=0.1 ~ mean 10 yrs)}

\item{beta0}{intercept for logistic model for false negative fraction}

\item{beta1}{slope of log(yi) for logistic model for false negative fraction}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}
}
\value{
vector of likelihoods
}
\description{
Do likelihood calculations for ScreeningModel2 using MVK onset and Exponential Sojourn
}
