% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScreeningModel3MVKLognorm}
\alias{ScreeningModel3MVKLognorm}
\alias{screening_model_3_likes_MVK_lognorm}
\title{Do likelihood calculations for ScreeningModel3 using MVK onset and Lognormal Sojourn}
\usage{
screening_model_3_likes_MVK_lognorm(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  mulog = 2.3,
  sdlog = 0.6,
  beta0 = -3,
  beta1 = 1,
  PrFalseNegBx = 0.05,
  tol = 1e-06,
  return_type = "",
  weights = NULL,
  left_trunc = FALSE,
  incidence = NULL
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{A}{MVK parameter A (typically negative, related to net proliferation)}

\item{B}{MVK parameter B (typically positive, related to malignant transformation)}

\item{delta}{MVK parameter delta (ratio of initiation rate to cell division rate)}

\item{mulog}{Lognormal meanlog for clinical diagnosis (Sojourn density, default=2.3 ~ 10 yrs)}

\item{sdlog}{Lognormal sdlog for clinical diagnosis (Sojourn density, default=0.6)}

\item{beta0}{intercept for logistic model for false negative fraction}

\item{beta1}{slope of log(yi) for logistic model for false negative fraction}

\item{PrFalseNegBx}{probability of a false negative biopsy | cancer, biopsy undertaken}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}

\item{return_type}{string, if "weighted_ll" returns sum of weighted log-likelihoods (default "")}

\item{weights}{vector of weights corresponding to inputs, required if return_type is "weighted_ll"}

\item{left_trunc}{bool, apply left truncation adjustment (default false)}

\item{incidence}{DataFrame containing background incidence rates, required if left_trunc is true}
}
\value{
vector of likelihoods (or vector of length 1 containing weighted log-likelihood sum)
}
\description{
Do likelihood calculations for ScreeningModel3 using MVK onset and Lognormal Sojourn
}
