% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScreeningModel4LikesLoglin}
\alias{ScreeningModel4LikesLoglin}
\alias{screening_model_4_likes_loglin}
\title{Do likelihood calculations for ScreeningModel4 using MVK onset, Exponential Sojourn, and Log-linear PSA with change of slope for onset}
\usage{
screening_model_4_likes_loglin(
  inputs,
  A = -0.1,
  B = 1e-04,
  delta = 1e-04,
  rate = 0.1,
  beta0 = 3.2892,
  beta1 = -0.5533,
  b0_psa = -1.6094,
  b1_psa = 0.02,
  b2_psa = 0.1094,
  sigma_psa = 0.2879,
  PrFalseNegBx = 0,
  tol = 1e-06,
  return_type = "",
  weights = NULL,
  left_trunc = FALSE,
  incidence = NULL
)
}
\arguments{
\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times, yi for biomarker values, bxi for biopsy indicators, and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{A}{MVK parameter A (typically negative, related to net proliferation)}

\item{B}{MVK parameter B (typically positive, related to malignant transformation)}

\item{delta}{MVK parameter delta (ratio of initiation rate to cell division rate)}

\item{rate}{Exponential rate for clinical diagnosis (Sojourn density, default=0.1 ~ mean 10 yrs)}

\item{beta0}{intercept for logistic model for false negative fraction}

\item{beta1}{slope of log(yi) for logistic model for false negative fraction}

\item{b0_psa}{intercept for log-linear PSA model}

\item{b1_psa}{slope for age for log-linear PSA model}

\item{b2_psa}{slope increment after onset for log-linear PSA model}

\item{sigma_psa}{standard deviation of log(PSA)}

\item{PrFalseNegBx}{probability of a false negative biopsy | cancer, biopsy undertaken}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}

\item{return_type}{string, if "weighted_ll" returns sum of weighted log-likelihoods (default "")}

\item{weights}{vector of weights corresponding to inputs, required if return_type is "weighted_ll"}

\item{left_trunc}{bool, apply left truncation adjustment (default false)}

\item{incidence}{DataFrame containing background incidence rates, required if left_trunc is true}
}
\value{
vector of likelihoods (or vector of length 1 containing weighted log-likelihood sum)
}
\description{
Do likelihood calculations for ScreeningModel4 using MVK onset, Exponential Sojourn, and Log-linear PSA with change of slope for onset
}
