% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\docType{package}
\name{screening}
\alias{screening-package}
\alias{screening}
\title{screening: Implement various screening models}
\description{
This currently implements the Day and Walter models.
}
\keyword{internal}
