% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScreeningModel3}
\alias{ScreeningModel3}
\alias{screening_model_3_predictions}
\alias{screening_model_3_likes}
\title{Do predictions for ScreeningModel3}
\usage{
screening_model_3_predictions(
  t,
  ti,
  yi,
  bxi,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta0 = -3,
  beta1 = 1,
  PrFalseNegBx = 0.05,
  simple = TRUE,
  tol = 1e-06
)

screening_model_3_likes(
  inputs,
  shape1 = 1,
  scale1 = 1,
  shape2 = 1,
  scale2 = 1,
  beta0 = -3,
  beta1 = 1,
  PrFalseNegBx = 0.05,
  tol = 1e-06,
  return_type = "",
  weights = NULL
)
}
\arguments{
\item{t}{double vector of times to evaluate}

\item{ti}{double vector of screening times}

\item{yi}{double vector of screening times}

\item{bxi}{integer vector of biopsy choices}

\item{shape1}{Weibull shape for onset}

\item{scale1}{Weibull scale for onset}

\item{shape2}{Weibull shape for clinical diagnosis}

\item{scale2}{Weibull scale for clinical diagnosis}

\item{beta0}{intercept for logistic model for false negative fraction}

\item{beta1}{slope of log(yi) for logistic model for false negative fraction}

\item{PrFalseNegBx}{probability of a false negative biopsy | cancer, biopsy undertaken}

\item{simple}{whether to calculate Z using a simple method (complement) or integration (defaults to true)}

\item{tol}{double for the numeric tolerance of the integration (default=1e-6)}

\item{inputs}{list of list with elements of t for the evaluation time, tj for the screening times and type for the type of likelihood (1=No cancer detected, 2=Screen-detected cancer, 3=Interval cancer)}

\item{return_type}{string, if "weighted_ll" returns sum of weighted log-likelihoods (default "")}

\item{weights}{vector of weights corresponding to inputs, required if return_type is "weighted_ll"}
}
\value{
data-frame with elements t, X, Y, Z (for state probabilities) and I (for incidence)

vector of likelihoods (or vector of length 1 containing weighted log-likelihood sum)
}
\description{
Do predictions for ScreeningModel3

Do likelihood calculations for ScreeningModel3
}
\examples{
par(mfrow=1:2)
x=seq(0,30,length=301)
screening_model_3_predictions(t=x, ti=vector("double"), yi=vector("double"), bxi=vector("integer"), shape1=1.5, shape2=1.5, scale1=20, scale2=10) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="No screening"))
screening_model_3_predictions(t=x, ti=c(10,15, 20, 25), yi=c(3,3,3), bxi=c(FALSE,FALSE,TRUE), shape1=1.5, shape2=1.5, scale1=20, scale2=10) |>
    with(matplot(time,cbind(X,Y,Z),type="l", lty=1, main="Screening"))
}
